added a few things, checkout README
