+++ README.md
@@ -1,2 +1,6 @@
 # Match_Game1
-New Updates
+
+## Setting up
+- Install Node.JS
+- In a terminal, run ``` npm i ``` to set things up
+- To start, simply run ``` npm run serve ```, will start a server a watch for code changes, perform static analysis and let you know if you do crappy stuff.
