+++ src/client/css/app.css
@@ -8,7 +8,7 @@ div#memory_board{
 	margin:0px auto;
 }
 div#memory_board > div{
-	background: url(img.png) no-repeat;
+	background: url('/assets/img.png') no-repeat;
 	border:#000 1px solid;
 	width:71px;
 	height:71px;
